Fix : could throw an exception on some async loads
If the loadContentToDisplay() was triggering a refresh synchronously (started loading for example), a flutter exception was triggered : we are not allowed to trigger a refresh while we are still building.
We now wrap the call to loadContentToDisplay() in a Future to make it wait for the build to finish before being called.